backend/requirements.txt
Django==4.2.9
python-dotenv==1.1.1
django-debug-toolbar==6.0.0